fix: pnpm recursive w/o subcmd
