+++ packages/plugin-commands-installation/src/recursive.ts
@@ -1,17 +1,13 @@
 import {
   createLatestSpecs,
-  docsUrl,
   getPinnedVersion,
   getSaveType,
   RecursiveSummary,
   throwOnCommandFail,
   updateToLatestSpecsFromManifest,
 } from '@pnpm/cli-utils'
-import { FILTERING } from '@pnpm/common-cli-options-help'
-import { Config, types as allTypes, WsPkg, WsPkgsGraph } from '@pnpm/config'
-import { WANTED_LOCKFILE } from '@pnpm/constants'
+import { Config, WsPkg, WsPkgsGraph } from '@pnpm/config'
 import { scopeLogger } from '@pnpm/core-loggers'
-import PnpmError from '@pnpm/error'
 import { arrayOfWorkspacePackagesToMap } from '@pnpm/find-workspace-packages'
 import logger from '@pnpm/logger'
 import { rebuild } from '@pnpm/plugin-commands-rebuild'
@@ -20,7 +16,6 @@ import sortPackages from '@pnpm/sort-packages'
 import { createOrConnectStoreController, CreateStoreControllerOptions } from '@pnpm/store-connection-manager'
 import { ImporterManifest, PackageManifest } from '@pnpm/types'
 import camelcaseKeys = require('camelcase-keys')
-import { oneLine } from 'common-tags'
 import isSubdir = require('is-subdir')
 import mem = require('mem')
 import fs = require('mz/fs')
@@ -29,7 +24,6 @@ import pLimit from 'p-limit'
 import path = require('path')
 import R = require('ramda')
 import readIniFile = require('read-ini-file')
-import renderHelp = require('render-help')
 import {
   addDependenciesToPackage,
   install,
@@ -38,133 +32,6 @@ import {
   mutateModules,
 } from 'supi'
 import { createWorkspaceSpecs, updateToWorkspacePackagesFromManifest } from './updateWorkspaceDependencies'
-
-export const rcOptionsTypes = cliOptionsTypes
-
-export function cliOptionsTypes () {
-  return {
-    access: ['public', 'restricted'],
-    recursive: Boolean,
-    table: Boolean,
-    ...R.pick([
-      'bail',
-      'link-workspace-packages',
-      'reporter',
-      'shared-workspace-lockfile',
-      'sort',
-      'tag',
-      'workspace-concurrency',
-    ], allTypes),
-  }
-}
-
-export function help () {
-  return renderHelp({
-    description: oneLine`
-      Concurrently performs some actions in all subdirectories with a \`package.json\` (excluding node_modules).
-      A \`pnpm-workspace.yaml\` file may be used to control what directories are searched for packages.`,
-    descriptionLists: [
-      {
-        title: 'Commands',
-
-        list: [
-          {
-            name: 'install',
-          },
-          {
-            name: 'add',
-          },
-          {
-            name: 'update',
-          },
-          {
-            description: 'Uninstall a dependency from each package',
-            name: 'remove <pkg>...',
-          },
-          {
-            description: 'Removes links to local packages and reinstalls them from the registry.',
-            name: 'unlink',
-          },
-          {
-            description: 'List dependencies in each package.',
-            name: 'list [<pkg>...]',
-          },
-          {
-            description: 'List packages that depend on <pkg>.',
-            name: 'why <pkg>...',
-          },
-          {
-            description: 'Check for outdated dependencies in every package.',
-            name: 'outdated [<pkg>...]',
-          },
-          {
-            description: oneLine`
-              This runs an arbitrary command from each package's "scripts" object.
-              If a package doesn't have the command, it is skipped.
-              If none of the packages have the command, the command fails.`,
-            name: 'run <command> [-- <args>...]',
-          },
-          {
-            description: `This runs each package's "test" script, if one was provided.`,
-            name: 'test [-- <args>...]',
-          },
-          {
-            description: oneLine`
-              This command runs the "npm build" command on each package.
-              This is useful when you install a new version of node,
-              and must recompile all your C++ addons with the new binary.`,
-            name: 'rebuild [[<@scope>/<name>]...]',
-          },
-          {
-            description: `Run a command in each package.`,
-            name: 'exec -- <command> [args...]',
-          },
-          {
-            description: 'Publishes packages to the npm registry. Only publishes a package if its version is not taken in the registry.',
-            name: 'publish [--tag <tag>] [--access <public|restricted>]',
-          },
-        ],
-      },
-      {
-        title: 'Options',
-
-        list: [
-          {
-            description: 'Continues executing other tasks even if a task threw an error.',
-            name: '--no-bail',
-          },
-          {
-            description: 'Set the maximum number of concurrency. Default is 4. For unlimited concurrency use Infinity.',
-            name: '--workspace-concurrency <number>',
-          },
-          {
-            description: oneLine`
-              Locally available packages are linked to node_modules instead of being downloaded from the registry.
-              Convenient to use in a multi-package repository.`,
-            name: '--link-workspace-packages',
-          },
-          {
-            description: 'Sort packages topologically (dependencies before dependents). Pass --no-sort to disable.',
-            name: '--sort',
-          },
-          {
-            description: oneLine`
-              Creates a single ${WANTED_LOCKFILE} file in the root of the workspace.
-              A shared lockfile also means that all dependencies of all workspace packages will be in a single node_modules.`,
-            name: '--shared-workspace-lockfile',
-          },
-        ],
-      },
-      FILTERING,
-    ],
-    url: docsUrl('recursive'),
-    usages: [
-      'pnpm recursive [command] [flags] [--filter <package selector>]',
-      'pnpm multi [command] [flags] [--filter <package selector>]',
-      'pnpm m [command] [flags] [--filter <package selector>]'
-    ],
-  })
-}
 
 type RecursiveOptions = CreateStoreControllerOptions & Pick<Config,
   'bail' |
